tests: fixing more harness issues with fetch
