+++ ui/test/setup.ts
@@ -7,6 +7,14 @@ declare global {
 globalThis.IS_REACT_ACT_ENVIRONMENT = true;
 
 import '@testing-library/jest-dom';
+import { rest } from 'msw';
+import { setupServer } from 'msw/node';
+import fetch from 'cross-fetch';
+
+vi.stubGlobal('fetch', (url: string, init: any) => {
+  // console.log(url, init);
+  return fetch(`http://localhost:3000${url}`, init);
+});
 
 // Prevent vite from failing when resizeObserver is used
 
@@ -42,3 +50,54 @@ Object.defineProperty(window, 'matchMedia', {
     dispatchEvent: vi.fn(),
   })),
 });
+
+const server = setupServer(
+  // rest.all('*', async (req, res, ctx) => {
+  //   const body = await req.json();
+  //   console.log(req, body);
+  // }),
+  rest.put('http://localhost:3000/~/channel/*', (req, res, ctx) => {
+    return res(
+      ctx.set('Connection', 'keep-alive'),
+      ctx.status(204),
+      ctx.body('')
+    );
+  }),
+  rest.get('http://localhost:3000/~/channel/*', async (req, res, ctx) => {
+    // console.log('event stream', req);
+    return res(
+      ctx.status(200),
+      ctx.set('Connection', 'keep-alive'),
+      ctx.set('Content-Type', 'text/event-stream'),
+      ctx.body(JSON.stringify('hello'))
+    );
+  }),
+  rest.get('http://localhost:3000/~/scry/chat/draft/*', (req, res, ctx) =>
+    // respond using a mocked JSON body
+    res(
+      ctx.json({
+        whom: '~zod/test',
+        story: {
+          inline: ['test'],
+          block: [],
+        },
+      })
+    )
+  )
+);
+
+server.listen({
+  onUnhandledRequest: 'warn',
+});
+
+// establish API mocking before all tests
+beforeAll(() =>
+  server.listen({
+    onUnhandledRequest: 'warn',
+  })
+);
+// reset any request handlers that are declared as a part of our tests
+// (i.e. for testing one-time error scenarios)
+afterEach(() => server.resetHandlers());
+// clean up once the tests are done
+afterAll(() => server.close());
